加上手機版選擇畢業組別按鈕 (#10)
