+++ src/Pages/Students/Graduation/OverViewCard/Card/index.js
@@ -16,16 +16,9 @@ import CloseIcon from '@material-ui/icons/Close'
 import Grow from '@material-ui/core/Grow'
 import AnimatedProgress from '../../../../../Components/AnimatedProgress'
 import Dialog from '@material-ui/core/Dialog'
-import Button from '@material-ui/core/Button'
-import MenuItem from '@material-ui/core/MenuItem'
-import Menu from '@material-ui/core/Menu'
 import CourseList from './CourseList'
 import GeneralCourseList from './GeneralCourseList'
 import GeneralNewCourseList from './GeneralNewCourseList'
-import {
-  getGraduationInfo,
-  getGraduationInfoAssistantVersion
-} from '../../../../../Redux/Students/Actions/Graduation'
 import './style.css'
 
 const styles = theme => ({
@@ -48,10 +41,6 @@ const styles = theme => ({
   },
   progress: {
     backgroundColor: '#00a152'
-  },
-  root: {
-    fontFamily: 'Noto Sans CJK TC',
-    letterSpacing: 1
   }
 })
 
@@ -88,13 +77,9 @@ class Index extends React.Component {
     this.handleOpen = this.handleOpen.bind(this)
     this.handleClose = this.handleClose.bind(this)
     this.handleChange = this.handleChange.bind(this)
-    this.professionalMenuOpen = this.professionalMenuOpen.bind(this)
-    this.professionalMenuClose = this.professionalMenuClose.bind(this)
-    this.professionalMenuSelect = this.professionalMenuSelect.bind(this)
     this.state = {
       open: false,
-      expanded: true,
-      anchorEl: null
+      expanded: true
     }
   }
 
@@ -112,34 +97,9 @@ class Index extends React.Component {
     this.setState({ expanded: !this.state.expanded })
   }
 
-  professionalMenuOpen (event) {
-    event.stopPropagation()
-    this.setState({
-      anchorEl: event.currentTarget
-    })
-  }
-
-  professionalMenuClose (event) {
-    event.stopPropagation()
-    this.setState({
-      anchorEl: null
-    })
-  }
-
-  professionalMenuSelect (field, event) {
-    event.stopPropagation()
-    let idCard = this.props.idCard
-    if (this.props.assis) {
-      this.props.getGraduationInfoAssistantVersion(idCard.id, idCard.sname, idCard.program, field)
-    }
-    else {
-      this.props.getGraduationInfo({ professional_field: field })
-    }
-  }
-
   render () {
     const { classes, rwd } = this.props
-    const professionalGroup = ['網多組(網)', '網多組(多)', '資工組', '資電組']
+
     if (this.props.data === undefined) return ''
 
     // for mobile
@@ -226,44 +186,6 @@ class Index extends React.Component {
                   />
                 </div>
               </div>
-              {
-                this.props.title === '共同必修' &&
-                <div className='col-md-3' style={{ marginLeft: '-7%' }}>
-                  <Button
-                    variant='outlined'
-                    aria-owns={this.state.anchorEl ? 'simple-menu' : undefined}
-                    aria-haspopup='true'
-                    onClick={this.professionalMenuOpen}
-                    disabled={this.props.reviewCheck !== 0}
-                  >
-                    {
-                      !isNaN(this.props.professionalField)
-                        ? professionalGroup[this.props.professionalField]
-                        : '畢業組別選擇'
-                    }
-                  </Button>
-                  <Menu
-                    id='simple-menu'
-                    anchorEl={this.state.anchorEl}
-                    open={Boolean(this.state.anchorEl)}
-                    onClose={this.professionalMenuClose}
-                    className={classes.root}
-                  >
-                    <MenuItem className={classes.root} onClick={(e) => this.professionalMenuSelect(0, e)}>
-                      網多組(網)
-                    </MenuItem>
-                    <MenuItem className={classes.root} onClick={(e) => this.professionalMenuSelect(1, e)}>
-                      網多組(多)
-                    </MenuItem>
-                    <MenuItem className={classes.root} onClick={(e) => this.professionalMenuSelect(2, e)}>
-                      資工組
-                    </MenuItem>
-                    <MenuItem className={classes.root} onClick={(e) => this.professionalMenuSelect(3, e)}>
-                      資電組
-                    </MenuItem>
-                  </Menu>
-                </div>
-              }
             </ExpansionPanelSummary>
             <ExpansionPanelDetails>
               {
@@ -300,16 +222,10 @@ Index.propTypes = {
 
 const mapStateToProps = (state, ownProps) => ({
   data: state.Student.Graduation.detail.data.filter(t => t.title === ownProps.title)[0],
-  studentIdcard: state.Student.User.studentIdcard,
-  idCard: state.Student.Graduation.assistant.idCard,
-  assis: state.Student.Graduation.assistant.using,
-  reviewCheck: state.Student.Graduation.getReview.check,
-  professionalField: state.Student.Graduation.getReview.professionalField
+  assis: state.Student.Graduation.assistant.using
 })
 
 const mapDispatchToProps = (dispatch) => ({
-  getGraduationInfo: (payload) => dispatch(getGraduationInfo(payload)),
-  getGraduationInfoAssistantVersion: (id, sname, program, field) => dispatch(getGraduationInfoAssistantVersion(id, sname, program, field))
 })
 
 export default connect(mapStateToProps, mapDispatchToProps)(withStyles(styles)(Index))
